+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -166,6 +166,11 @@ New checks
   Finds potentially erroneous calls to ``reset`` method on smart pointers when
   the pointee type also has a ``reset`` method.
 
+- New :doc:`readability-use-cpp-style-comments
+  <clang-tidy/checks/readability/use-cpp-style-comments>` check.
+
+  Replaces C-style comments with C++-style comments.
+  
 New check aliases
 ^^^^^^^^^^^^^^^^^